Tweaked the Create Issue page
The Create Issue page now only allows the creator to assign the Issue
to a Version that has not been released. This cuts down on UI clutter
and helps alleviate the accidental assigning of new Issues to released
Versions. (The Edit Issue page still allows the selection of released
Versions.)